+++ extensions/omni_httpd/CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 * HTTP response should contain correct status code reasons [#781](https://github.com/omnigres/omnigres/pull/781)
+* Rejected WebSocket requests are sent an HTTP response [#797](https://github.com/omnigres/omnigres/pull/797)
 
 ### Changed