+++ extensions/omni_httpd/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - 2025-02-20
+
+### Fixed
+
+* Support for the upcoming Postgres 18 [#805](https://github.com/omnigres/omnigres/pull/805)
+
 ## [0.3.0] - 2025-02-04
 
 ### Added
@@ -149,3 +155,5 @@ Initial release following a few months of iterative development.
 [0.2.10]: [https://github.com/omnigres/omnigres/pull/763]
 
 [0.3.0]: [https://github.com/omnigres/omnigres/pull/713]
+
+[0.3.1]: [https://github.com/omnigres/omnigres/pull/805]